Extract regional settings from overmap
